template_service/requirements.txt
amqp==5.3.1
asgiref==3.10.0
billiard==4.2.2
celery==5.5.3
click==8.3.0
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
Django==5.2.8
djangorestframework==3.16.1
drf-yasg==1.21.11
gunicorn==23.0.0
inflection==0.5.1
kombu==5.5.4
packaging==25.0
pika==1.3.2
prompt_toolkit==3.0.52
psycopg2-binary==2.9.11
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
PyYAML==6.0.3
redis==7.0.1
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
uritemplate==4.2.0
vine==5.1.0
wcwidth==0.2.14
